Fix F32toi16b

+++ sdr/conversion.go
@@ -31,11 +31,9 @@ func ui8toc64(input []byte, output []complex64) {
 
 func F32toi16b(input []float32, output []byte, scale float32)
 
-// var F32toi16b = f32toi16b
-
 func f32toi16b(input []float32, output []byte, scale float32) {
 	for i := 0; i < len(input); i++ {
-		v := uint16(input[i] * scale)
+		v := uint16(int16(input[i] * scale))
 		output[i*2] = uint8(v & 0xff)
 		output[i*2+1] = uint8(v >> 8)
 	}
